requirements.txt
# --- Core AI Framework (LangChain 0.3+) ---
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
openai>=1.50.0
tiktoken>=0.7.0

# --- API Server (The "Brain" Interface) ---
fastapi>=0.115.0
uvicorn>=0.32.0

# --- Data Validation & Schema ---
pydantic>=2.9.0

# --- Ingestion Engine (YouTube/Video) ---
yt-dlp>=2024.10.0
youtube-transcript-api>=0.6.3
google-api-python-client>=2.127.0
isodate>=0.6.1

# --- Image Processing (Visual RAG) ---
opencv-python>=4.10.0
numpy>=1.26.0

# --- Utilities ---
python-dotenv>=1.0.1
pyyaml>=6.0.1
tenacity>=8.5.0

# --- Embedding (Local, Fast, FREE!) ---
# FastEmbed: Local embeddings with no API costs
# IMPORTANT (Windows): Requires Visual C++ Redistributable
#   Download: https://aka.ms/vs/17/release/vc_redist.x64.exe
#   Or run: .\install_windows_deps.ps1 (as Administrator)
fastembed>=0.3.0
onnxruntime>=1.16.0  # Required by FastEmbed for ONNX model inference

# --- Database (The Heavy Metal Stack) ---
sqlmodel>=0.0.16
psycopg2-binary>=2.9.9
pgvector>=0.2.0
asyncpg>=0.29.0

# --- PDF Processing (Phase 2.5 - Citadel Pipeline) ---
PyPDF2>=3.0.0
reportlab>=4.0.0